[mythtv-cmyth] Remove overlapping EPG entries and fill gaps
PVR updates EPG entries to remove overlaps and fills gaps by updating the
start time of entries. This is a problem for the addon to match entries with
the MythTV database using channelid and starttime.

The patch makes sure to remove gaps and overlapping EPG entries before
transferring them to XBMC.